requirements.txt
asgiref==3.9.1
boto3==1.40.31
botocore==1.40.31
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
django-storages==1.14.6
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
jmespath==1.0.1
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.10.0
